Finish amount.

+++ ProgramDesignHomework/data/purchase.h
@@ -12,9 +12,9 @@
 typedef struct Purchase {
   int component;
   uint64_t time;
-  uint16_t price;       // 进货单价
-  uint16_t quantity;    // 进货数量
-  uint64_t total;       //商品总价
+  int price;       // 进货单价
+  int quantity;    // 进货数量
+  int total;       //商品总价
   string retailer;  //商品销售商
 } Purchase;
